Allow specifying specific projects.
The code still defaults to looking at your watched projects configured
in gerrit.  This allows you to specify a different list of projects
to consider, instead.

I am set up to watch most projects in gerrit, but when looking for a
review, I usually only want to consider the projects for which I'm a
-core team member.  So, now if I only want to see nova changes, I can do
something like:

  $ next-review -u russellb openstack/nova openstack/python-novaclient