Drop pagination support
This changes the SSH query to only request a limit of 100 results, and
work with those confines. In subsequent patches, I'll add more behavior
to the query to make that sample size a more efficient selection than
just an arbitrary 100 by moving some of next-review's behaviors into the
query itself.

Closes #22